Add files via upload
The project is a census data analysis report for a modestly sized town, aimed at providing recommendations to the local government on what should be built on an unoccupied piece of land and what primary future investments should be made based on the population demographic. The report includes data cleaning, demographic analysis, and detailed analysis of religion and infirmity, employment status, and marital status. The analysis shows a high percentage of employed people, a low birth rate, and a healthy dependency ratio between adults and children. Christians are the majority religious group in the community, with a few growing religions, and infirmity is relatively low in the town. The report makes recommendations based on these findings.